# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OneStyle API",
    "description": "Express + PostgreSQL backend for OneStyle e-commerce app",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:5000", "description": "Local dev" },
    { "url": "https://your-render-backend.onrender.com", "description": "Production (Render)" }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "WishlistItem": {
        "type": "object",
        "required": ["product_id", "name", "category", "image_url", "new_price", "old_price"],
        "properties": {
          "product_id": { "type": "integer", "example": 5 },
          "name":       { "type": "string",  "example": "Chic Overlap Collar Blouse" },
          "category":   { "type": "string",  "enum": ["women", "men", "kid"] },
          "image_url":  { "type": "string",  "example": "product_5.png" },
          "new_price":  { "type": "number",  "format": "float", "example": 85.0 },
          "old_price":  { "type": "number",  "format": "float", "example": 120.5 }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string", "example": "Something went wrong." }
        }
      },
      "MessageResponse": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string", "example": "Success." }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/wishlist": {
      "get": {
        "operationId": "getWishlist",
        "summary": "Get the authenticated user's wishlist",
        "tags": ["Wishlist"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Wishlist items with product details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["wishlist"],
                  "properties": {
                    "wishlist": { "type": "array", "items": { "$ref": "#/components/schemas/WishlistItem" } }
                  }
                }
              }
            }
          },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "operationId": "clearWishlist",
        "summary": "Clear the entire wishlist (e.g. after checkout)",
        "tags": ["Wishlist"],
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": { "description": "Wishlist cleared", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
    "/api/wishlist/{productId}": {
      "post": {
        "operationId": "addToWishlist",
        "summary": "Add a product to the wishlist (idempotent)",
        "tags": ["Wishlist"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "productId", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 5 }
        ],
        "responses": {
          "200": { "description": "Added to wishlist", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "400": { "description": "Invalid productId", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Product not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      },
      "delete": {
        "operationId": "removeFromWishlist",
        "summary": "Remove a product from the wishlist",
        "tags": ["Wishlist"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          { "name": "productId", "in": "path", "required": true, "schema": { "type": "integer" }, "example": 5 }
        ],
        "responses": {
          "200": { "description": "Removed from wishlist", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/MessageResponse" } } } },
          "400": { "description": "Invalid productId", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "401": { "description": "Not authenticated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "404": { "description": "Item not in wishlist", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } },
          "500": { "description": "Internal server error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorResponse" } } } }
        }
      }
    },
...
  }
}